+++ trades/index.js
@@ -1,3 +1,5 @@
+const encodeTrade = require('./encode_trade');
 const manageTrades = require('./manage_trades');
+const serviceAnchoredTrades = require('./service_anchored_trades');
 
-module.exports = {manageTrades};
+module.exports = {encodeTrade, manageTrades, serviceAnchoredTrades};